Refactor: Alpha의 value 프로퍼티의 타입 Double -> Float 변경
ViewController의 alphaSliderValueChanged(_:) 액션 메서드에서 변경될 alpha값을 변환하는 과정을 줄이기 위해 Alpha의 value 프로퍼티 타입을 Float로 변경
그로 인한 다른 메서드들의 매개변수 타입 수정
Plane 구조체의 stepUpAlphaValue(), stepDownAlphaValue() 메서드에서 사용하는 newAlpha 상수의 값을 결정할때 Float 타입 값들끼리 뺄셈을 함으로 인해 소수의 값에 필요없는 값이 추가되어 해당 값 처리하는 과정을 추가

+++ DrawingApp/DrawingApp/Model/Alpha.swift
@@ -8,12 +8,12 @@
 import Foundation
 
 struct Alpha: Equatable {
-    enum OpacityLevel: Double, CaseIterable {
+    enum OpacityLevel: Float, CaseIterable {
         case one = 0.1, two = 0.2, three = 0.3, four = 0.4, five = 0.5
         case six = 0.6, seven = 0.7, eight = 0.8, nine = 0.9, ten = 1.0
     }
     let opacityLevel: OpacityLevel
-    let value: Double
+    let value: Float
     
     init(opacityLevel: OpacityLevel) {
         self.opacityLevel = opacityLevel
